agentic-rag-challenge/requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
pydantic==2.0.0
requests==2.31.0
beautifulsoup4==4.12.0
streamlit==1.28.0
openai==1.0.0
